+++ packages/transport-sse/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @spoosh/transport-sse
 
+## 0.1.2
+
+- Fix `crypto.randomUUID` error in non-secure contexts (HTTP)
+
 ## 0.1.1
 
 - Attach `resolveParser` and `resolveAccumulator` utilities to transport object under `utils` namespace